+++ test/vanilla/AcceptanceTests/test_required_optional.py
@@ -125,3 +125,13 @@ class TestRequiredOptional(object):
 
         with pytest.raises(ValidationError):
             client.explicit.post_required_class_property(None)
+
+    def test_models(self):
+        from requiredoptional.models import Error
+
+        if sys.version_info >= (3,5):
+            from requiredoptional.models._models_py3 import Error as ErrorPy3
+            assert Error == ErrorPy3
+        else:
+            from requiredoptional.models._models import Error as ErrorPy2
+            assert Error == ErrorPy2