+++ test/vanilla/AcceptanceTests/test_url.py
@@ -235,3 +235,13 @@ class TestUrl(object):
             None,
             None,
         )
+
+    def test_models(self):
+        from url.models import Error
+
+        if sys.version_info >= (3,5):
+            from url.models._models_py3 import Error as ErrorPy3
+            assert Error == ErrorPy3
+        else:
+            from url.models._models import Error as ErrorPy2
+            assert Error == ErrorPy2