+++ test/vanilla/AcceptanceTests/test_xml.py
@@ -173,3 +173,13 @@ class TestXml(object):
         assert len(acls) == 1
         assert acls[0].id == 'MTIzNDU2Nzg5MDEyMzQ1Njc4OTAxMjM0NTY3ODkwMTI='
         _assert_with_log(client.xml.put_acls, acls)
+
+    def test_models(self):
+        from xmlservice.models import Error
+
+        if sys.version_info >= (3,5):
+            from xmlservice.models._models_py3 import Error as ErrorPy3
+            assert Error == ErrorPy3
+        else:
+            from xmlservice.models._models import Error as ErrorPy2
+            assert Error == ErrorPy2